requirements.txt
Flask
nic-parser
tensorflow
tensorflow-hub
matplotlib
tensorflow_datasets
Pillow